Add utility to realize github markers for reference building. (#9)
Basically convert the indices into symbols to pass to build_single_reference.

Also added number_of_classic_markers to subset the pre-computed markers
to the same number as in the classic algorithm. But then I realized that the
pre-computed markers were already subsetted to the desired number, so
we don't need to do any more subsetting during realization.