Protect against None values when processing features. (#11)
This ignores any None values when building a mapping dictionary, treating them
as fundamentally incomparable (like R's NAs). The aim is to support missing
feature identifiers because sometimes we don't have valid identifiers when
mapping between feature types (e.g., going from Ensembl to symbols).

A consequence of this change is that match/factorize now return lists for the
indices, in order to preserve any Nones in the inputs. The caller is then
responsible for converting these to NumPy arrays (which will throw an error if
any None is present - correctly, as None has no valid conversion to an int).