Report only the first ID per feature in reference's row data.
This fixes the IDs for features with multiple Ensembl/Entrez IDs (due to the
fact that the references were mapped back from symbols).

We also provide an option to report multiple IDs per feature in the row data;
this gives callers the chance to resolve ambiguities manually. Whatever is
passed to the other functions still needs to be a list of strings, though.